Use cleaner way of routing and controller
