+++ app/views/admin/tags/show.html.erb
@@ -20,7 +20,7 @@
           <% @tag_moderators.each do |user| %>
               <li class="flex justify-between">
                 <a href="/<%= URL.user(user) %>">@<%= user.username %></a>
-                <%= form_with url: remove_tag_moderator_admin_tag_path(@tag.id), model: [:admin, @tag], method: :delete, data: { remote: false } do |f| %>
+                <%= form_with url: admin_tag_moderator_path(@tag.id), model: [:admin, @tag], method: :delete, data: { remote: false } do |f| %>
                   <%= f.hidden_field :user_id, value: user.id %>
                   <%= f.submit "Remove", class: "crayons-btn crayons-btn--danger crayons-btn--s" %>
                 <% end %>
@@ -32,7 +32,7 @@
           This tag currently has no moderators.
         </div>
       <% end %>
-      <%= form_with url: add_tag_moderator_admin_tag_path(@tag.id), model: [:admin, @tag], data: { remote: false } do |f| %>
+      <%= form_with url: admin_tag_moderator_path(@tag.id), model: [:admin, @tag], method: :post, data: { remote: false } do |f| %>
         <div class="crayons-field w-50">
           <%= f.label :user_id, "Add Moderator (ID):", class: "crayons-field__label" %>
           <%= f.text_field :user_id, class: "crayons-textfield" %>
